# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Groupe autostop",
    "description": "Projet qui vise à renseigner les utilisateurs sur des points d'autostop accessibles. L'utilisateur a la possibilité de visualiser les données avec différents formats: JSON, XML ou RDF/XML",
    "version": "1.0.0",
    "license": {
      "name": "M.I.T",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "externalDocs": {
    "description": "Find out more about our project",
    "url": "https://github.com/Sophie168-mel/groupe_api/"
  },
  "servers": [
    {
      "url": "https://groupmiashsm2api.herokuapp.com/"
    }
  ],
  "paths": {
    "/pouce": {
      "post": {
        "tags": [
          "Point"
        ],
        "summary": "Ajout des coordonnées GPS d'un nouvel emplacement d'autostop dans la base",
        "operationId": "addPoint",
        "requestBody": {
          "description": "Vérifier que le format de données en entrée est accepté par le serveur",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Point"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Insertion OK",
            "content": {}
          },
          "406": {
            "description": "Header Accept not acceptable",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/pouces": {
      "get": {
        "tags": [
          "Point"
        ],
        "summary": "Retourne les x premiers emplacements pour un territoire en paramètre",
        "description": "Renvoie les données aux formats JSON, XML, RDF/XML",
        "operationId": "limit",
        "parameters": [
          {
            "name": "echelle",
            "in": "query",
            "description": "Status values that need to be considered for filter",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "enum": [
                "code_commune",
                "code_departement",
                "code_region"
              ]
            }
          },
          {
            "name": "value",
            "in": "query",
            "description": "Status values that need to be considered for filter",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "enum": [
                "1033",
                "1",
                "84"
              ]
            }
          },
          {
            "name": "limite",
            "in": "query",
            "description": "Status values that need to be considered for filter",
            "required": false,
            "explode": true,
            "schema": {
              "type": "integer",
              "example": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Limite"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "xml": {
                    "name": "XZY"
                  },
                  "items": {
                    "$ref": "#/components/schemas/Limite"
                  }
                }
              },
              "application/rdf+xml": {
                "schema": {
                  "type": "array",
                  "xml": {
                    "name": "XZY"
                  },
                  "items": {
                    "$ref": "#/components/schemas/Limite"
                  }
                }
              }
            }
          },
          "406": {
            "description": "Header Accept not acceptable",
            "content": {}
          }
        }
      }
    },
    "/map": {
      "get": {
        "tags": [
          "Map"
        ],
        "summary": "Retourne tous les points gps des arrets d'autostop dans un rayon donné",
        "description": "Plusieurs autostops peuvent etre retournés dans un rayon définit par sa longitude et sa latitude",
        "operationId": "find_per_ray",
        "parameters": [
          {
            "name": "rayon",
            "in": "query",
            "description": "Entrer le rayon",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "center",
            "in": "query",
            "description": "Entrer la latitude puis la longitude (Système de projection: Lambdert 93). Exemple: 45.140195,5.673187",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "string",
                "example": "45.140195,5.673187"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "xml": {
                    "name": "XZY"
                  },
                  "items": {
                    "$ref": "#/components/schemas/Map"
                  }
                }
              },
              "application/rdf+xml": {
                "schema": {
                  "type": "array",
                  "xml": {
                    "name": "XZY"
                  },
                  "items": {
                    "$ref": "#/components/schemas/Map"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Map"
                  }
                }
              }
            }
          },
          "406": {
            "description": "Header Accept not acceptable",
            "content": {}
          }
        }
      }
    },
    "/map/count": {
      "get": {
        "tags": [
          "Map"
        ],
        "summary": "Retourne le nombre de points gps d'arrets d'autostop d'une échelle donnée",
        "operationId": "mapCount",
        "parameters": [
          {
            "name": "echelle",
            "in": "query",
            "description": "Département ou Région ou Commune",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "enum": [
                "code_commune",
                "code_region",
                "code_departement"
              ]
            }
          },
          {
            "name": "value",
            "in": "query",
            "description": "Status values that need to be considered for filter",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "example": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "xml": {
                    "name": "XZY"
                  },
                  "items": {
                    "$ref": "#/components/schemas/id_territ"
                  }
                }
              },
              "application/rdf+xml": {
                "schema": {
                  "type": "array",
                  "xml": {
                    "name": "XZY"
                  },
                  "items": {
                    "$ref": "#/components/schemas/id_territ"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/id_territ"
                  }
                }
              }
            }
          },
          "406": {
            "description": "Header Accept not acceptable",
            "content": {}
          }
        }
      }
    },
    "/map/data": {
      "get": {
        "tags": [
          "Map"
        ],
        "summary": "Retourne la description d'un point d'autostop pour un id donné",
        "operationId": "map_data",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Identifiant de territoire",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer",
              "example": "0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "xml": {
                    "name": "XZY"
                  },
                  "items": {
                    "$ref": "#/components/schemas/descript"
                  }
                }
              },
              "application/rdf+xml": {
                "schema": {
                  "type": "array",
                  "xml": {
                    "name": "XZY"
                  },
                  "items": {
                    "$ref": "#/components/schemas/descript"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/descript"
                  }
                }
              }
            }
          },
          "406": {
            "description": "Header Accept not acceptable",
            "content": {}
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Point": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "example": 46.11581
          },
          "longitude": {
            "type": "number",
            "example": 5.81598
          },
          "commentaire": {
            "type": "string",
            "example": "ajout POI"
          },
          "code_commune": {
            "type": "number",
            "example": 73000
          },
          "nom_commune": {
            "type": "string",
            "example": "Chambéry"
          },
          "code_departement": {
            "type": "number",
            "example": 73
          },
          "nom_departement": {
            "type": "string",
            "example": "Savoie"
          },
          "code_region": {
            "type": "number",
            "example": 84
          },
          "nom_region": {
            "type": "string",
            "example": "Auvergne-Rhône-Alpes"
          }
        },
        "xml": {
          "name": "Point"
        }
      },
      "Limite": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "geohash": {
            "type": "string",
            "example": "u0hnk75h44"
          },
          "latitude": {
            "type": "string",
            "example": "46.11581"
          },
          "nom_commune": {
            "type": "string",
            "example": "Bellegarde-sur-Valserine"
          },
          "code_region": {
            "type": "string",
            "example": "84"
          },
          "id": {
            "type": "string",
            "example": "0"
          },
          "nom_region": {
            "type": "string",
            "example": "Auvergne-Rhône-Alpes"
          },
          "nom_departement": {
            "type": "string",
            "example": "Ain"
          },
          "code_commune": {
            "type": "string",
            "example": "1033"
          },
          "longitude": {
            "type": "string",
            "example": "5.81598"
          },
          "commentaire": {
            "type": "string",
            "example": "Bellegarde Valséo"
          },
          "code_departement": {
            "type": "string",
            "example": "1"
          }
        },
        "xml": {
          "name": "point"
        }
      },
      "Map": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "longitude": {
            "type": "string",
            "example": "5.6347541"
          },
          "code_region": {
            "type": "string",
            "example": "84"
          },
          "id": {
            "type": "string",
            "example": "4111"
          },
          "nom_commune": {
            "type": "string",
            "example": "Saint-Paul-de-Varces"
          },
          "commentaire": {
            "type": "string",
            "example": "Lourme"
          },
          "geohash": {
            "type": "string",
            "example": "u0h02hpnek"
          },
          "nom_departement": {
            "type": "string",
            "example": "Isère"
          },
          "nom_region": {
            "type": "string",
            "example": "Auvergne-Rhône-Alpes"
          },
          "latitude": {
            "type": "string",
            "example": "45.0670596"
          },
          "code_commune": {
            "type": "string",
            "example": "38436"
          },
          "code_departement": {
            "type": "string",
            "example": "38"
          }
        },
        "xml": {
          "name": "Map"
        }
      },
      "id_territ": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "Nb_per_territ": {
            "type": "integer",
            "example": "50"
          }
        },
        "xml": {
          "name": "Nombre"
        }
      },
      "descript": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "longitude": {
            "type": "string",
            "example": "5.6347541"
          },
          "code_region": {
            "type": "string",
            "example": "84"
          },
          "id": {
            "type": "string",
            "example": "4111"
          },
          "nom_commune": {
            "type": "string",
            "example": "Saint-Paul-de-Varces"
          },
          "commentaire": {
            "type": "string",
            "example": "Lourme"
          },
          "geohash": {
            "type": "string",
            "example": "u0h02hpnek"
          },
          "nom_departement": {
            "type": "string",
            "example": "Isère"
          },
          "nom_region": {
            "type": "string",
            "example": "Auvergne-Rhône-Alpes"
          },
          "latitude": {
            "type": "string",
            "example": "45.0670596"
          },
          "code_commune": {
            "type": "string",
            "example": "38436"
          },
          "code_departement": {
            "type": "string",
            "example": "38"
          }
        },
        "xml": {
          "name": "Nombre"
        }
      }
    }
  }
}